Refactor from flat folders to hierarchy (??)
