+++ src/index.js
@@ -1,16 +1,16 @@
 import ghurl from 'github-url-to-object'
-import { init } from './commands/init'
-import { fetch } from './commands/fetch'
-import { checkout } from './commands/checkout.js'
-import { list } from './commands/list.js'
-import { add } from './commands/add.js'
-import { remove } from './commands/remove.js'
-import { commit } from './commands/commit.js'
-import { verify } from './commands/verify.js'
-import { pack } from './commands/pack-objects.js'
-import { push } from './commands/push.js'
-import { getConfig } from './commands/getConfig.js'
-import { setConfig } from './commands/setConfig.js'
+import { init } from './init'
+import { fetch } from './fetch'
+import { checkout } from './checkout.js'
+import { list } from './list.js'
+import { add } from './add.js'
+import { remove } from './remove.js'
+import { commit } from './commit.js'
+import { verify } from './verify.js'
+import { pack } from './pack-objects.js'
+import { push } from './push.js'
+import { getConfig } from './getConfig.js'
+import { setConfig } from './setConfig.js'
 
 export default function git (dir) {
   return new Git(dir)
