+++ src/managers/models/GitCommit.js
