+++ src/managers/models/GitConfig.js
