+++ src/managers/models/GitIndex.js
