+++ src/managers/models/utils/delete.js
