+++ src/managers/models/utils/exists.js
