+++ src/managers/models/utils/flatFileListToDirectoryStructure.js
