+++ src/managers/models/utils/pkt-line-encoder.js
