+++ src/pack-objects.js
@@ -1,6 +1,6 @@
 // @flow
 import { Buffer } from 'buffer'
-import GitObjectManager from '../managers/GitObjectManager'
+import GitObjectManager from './managers/GitObjectManager'
 import pad from 'pad'
 import pako from 'pako'
 import crypto from 'crypto'
