+++ src/setConfig.js
@@ -1,4 +1,4 @@
-import GitConfigManager from '../managers/GitConfigManager'
+import GitConfigManager from './managers/GitConfigManager'
 
 export async function setConfig ({ gitdir, path, value }) {
   const config = await GitConfigManager.get({ gitdir })
