+++ test/test-commit.js
@@ -1,7 +1,7 @@
 import test from 'ava'
 import git from '..'
-import { read } from '../lib/utils/read'
-import { write } from '../lib/utils/write'
+import { read } from '../lib/managers/models/utils/read'
+import { write } from '../lib/managers/models/utils/write'
 import jsonfile from 'jsonfile'
 import pify from 'pify'
 import path from 'omnipath'
